requirements.txt
aiohttp==3.9.5
aiolimiter==1.1.0
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anthropic==0.26.1
anyio==4.3.0
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
build==1.2.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
colorlog==6.8.2
crashtest==0.4.1
datasets==2.14.7
dill==0.3.7
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
dspy-ai==2.4.9
dulwich==0.21.7
fastjsonschema==2.19.1
filelock==3.14.0
frozenlist==1.4.1
fsspec==2023.10.0
google==3.0.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.1
idna==3.7
importlib_metadata==7.1.0
install==1.3.5
installer==0.7.0
jaraco.classes==3.4.0
jiter==0.1.0
joblib==1.3.2
jsonargparse==4.28.0
keyring==24.3.1
llm_judge==0.1
Mako==1.3.5
MarkupSafe==2.1.5
more-itertools==10.2.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.15
numpy==1.26.4
openai==1.30.1
optuna==3.6.1
packaging==24.0
pandas==2.2.2
pexpect==4.9.0
pkginfo==1.10.0
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==3.12.3
pyparsing==3.1.2
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.1
regex==2024.5.15
requests==2.32.1
requests-toolbelt==1.0.0
rsa==4.9
scikit-learn==1.5.0
scipy==1.13.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
structlog==24.1.0
tenacity==8.3.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomlkit==0.12.5
tqdm==4.66.4
trove-classifiers==2024.5.17
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.2
voyageai==0.2.2
xattr==1.1.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.2
